Fix to pass build

+++ domains/repositories/song_repository.go
@@ -3,6 +3,6 @@ package repositories
 import "github.com/Basabi-lab/lms/domains/models"
 
 type SongRepository interface {
-	GetByID(id int64) (*models.Song, error)
-	Create(song *models.Song) (int64, error)
+	GetByID(id uint) (*models.Song, error)
+	Create(song *models.Song) (uint, error)
 }
